Remove the "Line Ending Style" menu and switch to using a statusbar combo box.
Remove the "Line Ending Style" menu and instead add a `GeditStatusComboBox`
element to the statusbar for switching between line ending styles.

Many Gedit installations at this time do not expose the custom
`GeditStatusComboBox` widget to the Python `gedit` module, so this commit
adds a pure PyGTK port that is used whenever `GeditStatusComboBox` is not
natively available.